src/inference.py
"""
Inference engine for road damage detection
"""
from typing import List, Dict, Union
from pathlib import Path
import torch
from PIL import Image
import numpy as np


class RoadDamageDetector:
    """Main detector class for road damage detection"""

    def __init__(
        self,
        model_paths: List[str],
        mode: str = 'accurate',
        device: str = None
    ):
        """
        Initialize the road damage detector

        Args:
            model_paths: List of paths to model weights
            mode: Detection mode ('fast', 'balanced', 'accurate')
            device: Device to run inference on ('cuda' or 'cpu')
        """
        self.model_paths = model_paths
        self.mode = mode
        self.device = device or ('cuda' if torch.cuda.is_available() else 'cpu')

        # TODO: Implement model loading
        print(f"Initialized RoadDamageDetector with {len(model_paths)} models")
        print(f"Mode: {mode}, Device: {self.device}")

    def predict(self, image: Union[str, Image.Image]) -> List[Dict]:
        """
        Run inference on an image

        Args:
            image: Path to image or PIL Image

        Returns:
            List of detections with bbox, class, confidence
        """
        # TODO: Implement prediction logic
        return []

    def visualize(self, results: List[Dict], save_path: str = None):
        """
        Visualize detection results

        Args:
            results: Detection results
            save_path: Optional path to save visualization
        """
        # TODO: Implement visualization
        pass
